State updates
- change the 'LOAD' column to display unit_file_state when this field is
  present. When its not, revert to displaying value. This makes the
  column more useful as now it can show enablement.

- Make the service show/update enable/disabled/masked/unmasked state

- This was done earlier but forgot to mention. Make it so that all
  characters are converted to lower case before their action is
  inspected. This allows upper case characters to work the same way the
  lowers do. Given the instructions give capital letters this makes
  sense to do that.

Note: When systemctl does an enable/disable/mask/unmask it also issues a
reload. We dont do that because that causes a visible 'flash' to appear
to the screen. Systemd doesn't really care if you do a reload or not.

If you do this in systemctl you'll notice it updates in the
servicemaster display but emits a flash when it does so when it detects
teh reload and re-inspects all the services.

Systemd is aware of the enablement state of the unit (since it can be
queried directly which is how we do it).