Absolutely minimize screen blinking/blanking on update.
- Track a last_updated value to know when we last updated a service (got
  a message from dbus about a thing changing)
- Track changes inside of a service.
- On reload, no longer empty the service list and rebuild it, instead
  check if we have the service already and update it. This avoids
  clearing the screen on a reload.

Using these two parameters its possible to strategically erase only the
lines on scren to services that have recieved a change, either by user
inputs or via dbus.

The only time we erase the entire screen now is when a service has been
removed entirely from the system AND It was being displayed in the
current iteration of the screen.

The last updated value is touched with the latest event time whenever we
update a property or reload the service. On a reload, should a service
not receive an updated timestamp it means its not existing in the current
unit list in systemd, so we can remove it.