Major refactor/rewrite
Split the code into various modules.

- sm_err
Does error handling. Currently fairly basic. One that ends the curses
terminal, spits to stderr and exits.

The other hands off the error to a display window.

- services
Header file contains all the main service structure shared with other
modules along with functions meant for searching for services or
performing operations on groups of services. Also handles journal talk
for the service.

- bus
Provides header file for a bus and the bus object.
A bus in this sense contains the stateful data for that dbus service.
Tracking the total types in each systemd instance, its reloading state
and any services being ran.
Bus functions are dedicated to work done via dbus on behalf of service
objects. All the DBUS talk goes here.

- display
Providers headers and handlers for the curses related work and cursor
inputs. Used to display the services to the user and manage the display
state, such as whether to show user or system systemd instances.

- servicemaster.c
Provides the glue that puts all the objects together and is now fairly
small.