Cleanup error handling
Make sm_error a bit cleaner - initially this wasn't meant to exit and I
messed that up. It should probably be re-worked in that manner though.,

Cleaned up invocation id error handling. Some minor changes
- declare variables at of function (purely stylistic, happy for you to say no on
  this)
- You probably wanted an error to be displayed to the user, the
  "sm_error_window" will do this in the manner you're looking for.
- Dont return rc, but goto fin. This is because there is a cleanup
  routine at the end of the function that frees memory. If you just
  return you'll end up with a memory leak.

As an addage the huge rewrite *was* huge and I've basically made a lot
of internal API changes in regards to the naming scheme being used. So
no worries I made this a lot more complicated!

Note: At this point I've just tried to turn it into a 'mini project,
project' as a means to just understand the way I'm building C programs
(modules in separate files, etc). I realize that its not really that
necessary for a relatively small project like this.